moved category and article into management namespace. deleted unused route
