+++ app/controllers/orders_controller.rb
@@ -43,12 +43,12 @@ class OrdersController < ApplicationController
 
   def order_complete
     if @current_order.empty?
-      redirect_to categories_path, notice: 'Fehler: Keine Artikel im Warenkorb'
+      redirect_to management_categories_path, notice: 'Fehler: Keine Artikel im Warenkorb'
     elsif !@current_order.update(state: 'placed')
-      redirect_to categories_path, notice: 'Fehler: Bestellung konnte nicht aufgegeben werden'
+      redirect_to management_categories_path, notice: 'Fehler: Bestellung konnte nicht aufgegeben werden'
     else
       spawn_new_current_order
-      redirect_to categories_path, notice: 'Bestellung wurde abgeschlossen'
+      redirect_to management_categories_path, notice: 'Bestellung wurde abgeschlossen'
     end
   end
