--- app/views/articles/show.html.slim
@@ -1,6 +1,0 @@
-= render "article", article: @article, category: @category
-= render "articlecat", category: @category
-
-div
-  = link_to "Bearbeiten", edit_category_article_path(@category, @article)
-  = button_to "Löschen", category_article_path(@category, @article), method: :delete
