--- app/views/categories/index.html.slim
@@ -1,16 +1,0 @@
-h1
-  = "Kategorien"
-
-- @categories.sort_by { |cat| cat.name }.each do |category|
-  = render category
-  p
-    = link_to "Hier ... Artikel", category_articles_path(category)
-    - if user_signed_in?
-      = " |  "
-      = link_to "Editieren", edit_category_path(category)
-      = button_to "Löschen", category, method: :delete
-
-- if user_signed_in?
-  = link_to "Neue Artikelkategorie", new_category_path
-  = " | "
-  = link_to "Neuer Artikel", new_article_path
