--- app/views/categories/show.html.slim
@@ -1,7 +1,0 @@
-= render @category
-
-div
-  = link_to "Editieren", edit_category_path(@category)
-  = " | "
-  = link_to "Anzeigen", category_articles_path(@category)
-  = button_to "Löschen", @category, method: :delete
