+++ app/views/management/articles/_article.html.slim
