+++ app/views/management/articles/_form.html.slim
@@ -1,9 +1,9 @@
 ruby:
   if category
-    path = article.id ? category_article_path(category, article) : category_articles_path(category)
+    path = article.id ? management_category_article_path(category, article) : management_category_articles_path(category)
     http_method = article.id ? :patch : :post
   else
-    path = articles_path
+    path = management_articles_path
     http_method = :post
   end
