+++ app/views/management/articles/edit.html.slim
@@ -3,4 +3,4 @@ h1
 = render "form", category: @article.category, article: @article
 br
 div
-  = link_to "Anzeigen", category_article_path(@article.category)
+  = link_to "Anzeigen", management_category_article_path(@article.category)
