+++ app/views/management/articles/index.html.slim
@@ -7,7 +7,7 @@ h1
   - @articles.each do |article|
     = render "article", article: article
     p
-      = link_to "Anzeigen", category_article_path(@category, article)
+      = link_to "Anzeigen", management_category_article_path(@category, article)
       = " | "
       - line_item = LineItem.new(count: 1, article: article)
       = form_with(model: line_item) do |form|
@@ -15,4 +15,4 @@ h1
         = form.select :count, options_for_select(1..10)
         = form.submit "Zum Warenkorb hinzufügen"
 
-= link_to "Neuer Artikel", new_category_article_path(@category) if user_signed_in?
+= link_to "Neuer Artikel", new_management_category_article_path(@category) if user_signed_in?
