+++ app/views/management/articles/new.html.slim
