+++ app/views/management/categories/_category.html.slim
