+++ app/views/management/categories/edit.html.slim
@@ -2,4 +2,4 @@ h1= "Editing Kategorie"
 = render "form", category: @category
 br
 div
-  = link_to "Alle Kategorien", categories_path
+  = link_to "Alle Kategorien", management_categories_path
