+++ app/views/management/categories/show.html.slim
@@ -1,0 +1,7 @@
+= render @category
+
+div
+  = link_to "Editieren", edit_management_category_path(@category)
+  = " | "
+  = link_to "Anzeigen", management_category_articles_path(@category)
+  = button_to "Löschen", management_category_path(@category), method: :delete
